+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Added
+
+- Added a new named export `AuthTokenRefresh`. This is a
+  constructor that can be invoked to create a distinct instance, for
+  applications that require more than one Passport instance.
+
 ## [2.1.0] - 2021-06-29
 
 ### Added